refactor(app): use camelCase convention to file names
