+++ src/app/constants/gameConstants.ts
