+++ src/app/models/card.models/abstractCard.model.ts
