--- src/app/models/card.models/base-card.model.ts
@@ -1,3 +1,0 @@
-import { Card } from "./abstract-card.model";
-
-export class BaseCard extends Card {}
