+++ src/app/models/card.models/baseCard.model.ts
@@ -1,0 +1,3 @@
+import { Card } from "./abstractCard.model";
+
+export class BaseCard extends Card {}
