--- src/app/models/card.models/fate-card.model.ts
@@ -1,3 +1,0 @@
-import { Card } from "./abstract-card.model"
-
-export class FateCard extends Card {}
