+++ src/app/models/card.models/fateCard.model.ts
@@ -1,0 +1,3 @@
+import { Card } from "./abstractCard.model"
+
+export class FateCard extends Card {}
