--- src/app/models/card.models/hobbit-card.model.ts
@@ -1,3 +1,0 @@
-import { ObtainableCard } from "./abstract-card.model";
-
-export class HobbitCard extends ObtainableCard {}
