+++ src/app/models/card.models/hobbitCard.model.ts
@@ -1,0 +1,3 @@
+import { ObtainableCard } from "./abstractCard.model";
+
+export class HobbitCard extends ObtainableCard {}
