--- src/app/models/card.models/trap-card.model.ts
@@ -1,3 +1,0 @@
-import { Card } from "./abstract-card.model";
-
-export class TrapCard extends Card {}
