+++ src/app/models/card.models/trapCard.model.ts
@@ -1,0 +1,3 @@
+import { Card } from "./abstractCard.model";
+
+export class TrapCard extends Card {}
