+++ src/app/services/card.services/artifactCard.service.ts
@@ -1,7 +1,7 @@
 import { Player } from "../../models/player.model";
-import { ArtifactCard } from "../../models/card.models/artifact-card.model";
-import { playerOwnsCard, findCardOwner, numberOfOwnedCards, payMoneyToCardOwner} from "./shared-card.service";
-import { ARTIFACT_CARDS_IDS } from "../../constants/game-constants";
+import { ArtifactCard } from "../../models/card.models/artifactCard.model";
+import { playerOwnsCard, findCardOwner, numberOfOwnedCards, payMoneyToCardOwner} from "./sharedCard.service";
+import { ARTIFACT_CARDS_IDS } from "../../constants/gameConstants";
 
 export const payArtifactTax = (player: Player, players: Player[], artifact: ArtifactCard): void => {
   if(artifact.isObtainable) return;
