+++ src/app/services/card.services/trapCard.service.ts
@@ -1,8 +1,8 @@
 import { Player } from "../../models/player.model";
-import { playerLosesMoney } from "./shared-card.service";
+import { playerLosesMoney } from "./sharedCard.service";
 import { Messages, logMessage } from "../messasges.service";
 import { TRAP_TAX } from "../../constants/prices";
-import { PLAYER_IS_ATTACKED_STEPS, PLAYER_IS_CHASED_STEPS } from "../../constants/game-constants";
+import { PLAYER_IS_ATTACKED_STEPS, PLAYER_IS_CHASED_STEPS } from "../../constants/gameConstants";
 
 export const loseMoneyAndMoveBackward = (player: Player): void => {
   loseMoneyAndMove(player, TRAP_TAX, PLAYER_IS_CHASED_STEPS);
